# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "description": "User login",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "description": "Refresh token",
                "operationId": "b2eae7a8032954d25b46c7ec6f489c68",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthRefreshTokenDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth_user": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "description": "Logout token",
                "operationId": "69281b12abb272c76871f19cb17ca563",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthRefreshTokenDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmptyResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/file/upload/{type}": {
            "post": {
                "tags": [
                    "File"
                ],
                "description": "Upload file",
                "operationId": "84bfca39f201315104e2265e4357824e",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "type: avatar"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FileResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth_user": []
                    }
                ]
            }
        },
        "/api/file/get/{type}/{filename}": {
            "get": {
                "tags": [
                    "File"
                ],
                "description": "Get file",
                "operationId": "09b46299c8aa208d70ad0f1a6d32f383",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "type: avatar"
                    },
                    {
                        "name": "filename",
                        "in": "path",
                        "description": "filename"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "auth_user": []
                    }
                ]
            }
        },
        "/api/file/delete/avatar": {
            "delete": {
                "tags": [
                    "File"
                ],
                "description": "Delete avatar",
                "operationId": "552a10f16fbd3b9527fcc8c3cffa001b",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmptyResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth_user": []
                    }
                ]
            }
        },
        "/api/user/create": {
            "post": {
                "tags": [
                    "User"
                ],
                "description": "Create new user",
                "operationId": "a286e80786abff27a4c441322a18d3e4",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/validate/{field}": {
            "post": {
                "tags": [
                    "User"
                ],
                "description": "Validate user field. Values: email, nickname",
                "operationId": "1837420d84eb7711a8395ebd6c233da9",
                "parameters": [
                    {
                        "name": "field",
                        "in": "path",
                        "description": "email|nickname"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ValidateUserFieldDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmptyResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/password/forget": {
            "post": {
                "tags": [
                    "User"
                ],
                "description": "Password forget link sending",
                "operationId": "b6ebbf956e62415d76490a50c05184df",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ForgetPasswordDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForgetPasswordResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/password/reset": {
            "post": {
                "tags": [
                    "User"
                ],
                "description": "Password restore",
                "operationId": "4d5f4e7d6a85f4f309163ec6472b666b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetPasswordDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForgetPasswordResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Get current user",
                "operationId": "22ea85303059d4f2d15556c1c9fd65d2",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth_user": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "description": "Update current user",
                "operationId": "fd4f39f7820d3f3cbc259c5b30fed830",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth_user": []
                    }
                ]
            }
        },
        "/api/user/password/change": {
            "post": {
                "tags": [
                    "User"
                ],
                "description": "Change restore",
                "operationId": "552949b446745542733ef11c816c41b7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangePasswordDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth_user": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "AuthDTO": {
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "AuthRefreshTokenDTO": {
                "properties": {
                    "fingerprint": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ChangePasswordDTO": {
                "properties": {
                    "currentPassword": {
                        "type": "string"
                    },
                    "newPassword": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "CreateUserDTO": {
                "properties": {
                    "nickname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ForgetPasswordDTO": {
                "properties": {
                    "email": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ResetPasswordDTO": {
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "token": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateUserDTO": {
                "properties": {
                    "nickname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "timezone": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ValidateUserFieldDTO": {
                "properties": {
                    "field": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "AuthResource": {
                "properties": {
                    "result": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "accessToken": {
                                    "type": "string"
                                },
                                "refreshToken": {
                                    "type": "string"
                                },
                                "fingerprint": {
                                    "type": "string"
                                },
                                "tokenType": {
                                    "type": "string"
                                },
                                "accessTokenExpiredIn": {
                                    "type": "integer"
                                },
                                "refreshTokenExpiredIn": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "status": {
                        "type": "integer"
                    },
                    "meta": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "EmptyResource": {
                "properties": {
                    "result": {
                        "type": "object"
                    },
                    "status": {
                        "type": "integer"
                    },
                    "meta": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "FileResource": {
                "properties": {
                    "result": {
                        "$ref": "#/components/schemas/FileResourceItem"
                    },
                    "status": {
                        "type": "integer"
                    },
                    "meta": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ForgetPasswordResource": {
                "properties": {
                    "result": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "status": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "status": {
                        "type": "integer"
                    },
                    "meta": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "FileResourceItem": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "filename": {
                        "type": "string"
                    },
                    "extension": {
                        "type": "string"
                    },
                    "path": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserResourceItem": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "nickname": {
                        "type": "string"
                    },
                    "timezone": {
                        "type": "string"
                    },
                    "avatar": {
                        "$ref": "#/components/schemas/FileResourceItem"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "properties": {
                    "result": {
                        "$ref": "#/components/schemas/UserResourceItem"
                    },
                    "status": {
                        "type": "integer"
                    },
                    "meta": {
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "auth_user": {
                "type": "apiKey",
                "description": "JWT auth (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}